configured mention-tracking logic and cleaned code structure
